KI-40 [tfjs] Remove casting of output data in Size operator

+++ inference/inference-tfjs/src/jsMain/kotlin/io.kinference.tfjs/operators/tensor/Size.kt
@@ -52,7 +52,7 @@ class SizeVer1(
 
     override suspend fun <D : ONNXData<*, *>> apply(contexts: Contexts<D>, inputs: List<TFJSTensor?>): List<TFJSTensor?> {
         val inputSize = inputs[0]!!.data.linearSize
-        val dataSize = NDArrayTFJS.intScalar(inputSize.toLong())
+        val dataSize = NDArrayTFJS.intScalar(inputSize)
         return listOf(dataSize.asTensor("size"))
     }
 }
